chore: split antv (#43765)
* chore: split antv

* chore: add fallback

* chore: code clean

* chore: fix lint

* chore: fix lint
